Point numeration
refactoring ParagraphStructure
